nitpick(front): one function per file
